add env variable check for USER_DATA_DIR
It's a tricky variable for this usecase
If it's set to something other than an empty string, it will break
parallel testing, but you can configure the browser so you get
sourcemaps on arbitrary sites.

If it's set to an empty string, that allows the parallel testing to take
place.