wip add tagger-worker project
The current problem with vite is that managing multiple inputs/outputs
for different formats (specifically like iife and umd) doesn't work out
of the box.

There may be a way to do it through rollup alone, but that's a whole
other ordeal.

Fortunately, NX is kind of built for this kind of thing.

For the time being, tagger-worker will be the project that manages
anything related to the service worker so that the extension can get one
iife file that doesn't have the breaking imported polyfills.